+++ Runtime/ScriptableState.cs
@@ -9,15 +9,14 @@ namespace loophouse.ScriptableStates
         [SerializeField] ScriptableAction[] _exitActions;
         [SerializeField] ScriptableAction[] _physicsActions; //to be run in fixed update
         [SerializeField] ScriptableAction[] _stateActions; //to be run in update
-        [SerializeField] StateTransition[] _transitions; //to be run in late update
 
-        public void Begin(ScriptableStatesComponent statesComponent)
+        public void Begin(StateComponent stateComponent)
         {
             foreach (var action in _entryActions)
             {
                 if (action)
                 {
-                    action.Act(statesComponent);
+                    action.Act(stateComponent);
                 }
                 else
                 {
@@ -26,13 +25,13 @@ namespace loophouse.ScriptableStates
             }
         }
 
-        public void End(ScriptableStatesComponent statesComponent)
+        public void End(StateComponent stateComponent)
         {
             foreach (var action in _exitActions)
             {
                 if (action)
                 {
-                    action.Act(statesComponent);
+                    action.Act(stateComponent);
                 }
                 else
                 {
@@ -41,13 +40,13 @@ namespace loophouse.ScriptableStates
             }
         }
 
-        public void UpdatePhysics(ScriptableStatesComponent statesComponent)
+        public void UpdatePhysics(StateComponent stateComponent)
         {
             foreach (var action in _physicsActions)
             {
                 if (action)
                 {
-                    action.Act(statesComponent);
+                    action.Act(stateComponent);
                 }
                 else
                 {
@@ -56,13 +55,13 @@ namespace loophouse.ScriptableStates
             }
         }
 
-        public void UpdateState(ScriptableStatesComponent statesComponent)
+        public void UpdateState(StateComponent stateComponent)
         {
             foreach (var action in _stateActions)
             {
                 if (action)
                 {
-                    action.Act(statesComponent);
+                    action.Act(stateComponent);
                 }
                 else
                 {
@@ -70,49 +69,5 @@ namespace loophouse.ScriptableStates
                 }
             }
         }
-
-        public ScriptableState CheckTransitions(ScriptableStatesComponent statesComponent, ScriptableState emptyState)
-        {
-            foreach (StateTransition transition in _transitions)
-            {
-                if (transition.condition)
-                {
-                    if (transition.condition.Verify(statesComponent))
-                    {
-                        if (transition.trueState != emptyState)
-                        {
-                            if (transition.trueState)
-                            {
-                                return transition.trueState;
-                            }
-                            else
-                            {
-                                Debug.LogError($"[SCRIPTABLE STATE] {name}'s Transitions list has an element with a null true state", this);
-                            }
-                        }
-                    }
-                    else
-                    {
-                        if (transition.falseState != emptyState)
-                        {
-                            if (transition.falseState)
-                            {
-                                return transition.falseState;
-                            }
-                            else
-                            {
-                                Debug.LogError($"[SCRIPTABLE STATE] {name}'s Transitions list has an element with a null false state", this);
-                            }
-                        }
-                    }
-                }
-                else
-                {
-                    Debug.LogError($"[SCRIPTABLE STATE] {name}'s Transitions list has an element with a null condition", this);
-                }
-            }
-            return emptyState;
-        }
-
     }
 }