+++ Runtime/StateTransition.cs
@@ -4,6 +4,7 @@ namespace loophouse.ScriptableStates
     [System.Serializable]
     public struct StateTransition
     {
+        public ScriptableState originState;
         public ScriptableCondition condition;
         public ScriptableState trueState;
         public ScriptableState falseState;