+++ Samples~/FollowBehaviourExample/Scripts/Actions/ConsumeStaminaAction.cs
@@ -6,7 +6,7 @@ using loophouse.ScriptableStates;
 [CreateAssetMenu(menuName = "Scriptable State Machine/Actions/ConsumeStaminaAction", fileName = "new ConsumeStaminaAction")]
 public class ConsumeStaminaAction : ScriptableAction
 {
-	public override void Act(ScriptableStatesComponent statesComponent)
+	public override void Act(StateComponent statesComponent)
 	{
 		statesComponent.GetComponent<StaminaComponent>().ConsumeStamina();
 	}